requirements.txt
opencv-python
Pillow
mysql-connector-python
numpy
face-recognition
